Fixed wrong type conversion in cstrLen() for debug mode as well
Apply the type conversion correction added by PR
https://github.com/rsyslog/rsyslog/pull/5051 to both definitions
of cstrLen().

Changed types from int to size_t in multiple occurrences
(where cstrLen is used).

closes: https://github.com/rsyslog/rsyslog/issues/5073

+++ parse.h
@@ -56,7 +56,7 @@ struct rsParsObject
 	rsObjID OID;			/**< object ID */
 #endif
 	cstr_t *pCStr;		/**< pointer to the string object we are parsing */
-	int iCurrPos;			/**< current parsing position (char offset) */
+	size_t iCurrPos;	/**< current parsing position (char offset) */
 };
 typedef struct rsParsObject rsParsObj;
